+++ packages/commercetools/theme/helpers/filters/applyFiltersFromUrl.js
@@ -1,0 +1,18 @@
+export const applyFiltersFromUrl = (context, filters) => {
+  const { query } = context.root.$route;
+  const filtersFromQuery = Object.entries(query).filter(([name]) => !['page', 'items'].includes(name));
+  filtersFromQuery.forEach(([name, values]) => {
+    if (!filters[name]) {
+      return;
+    }
+
+    if (!Array.isArray(values)) {
+      values = [values];
+    }
+    filters[name].options.forEach(option => {
+      if (values.includes(option.value)) {
+        option.selected = true;
+      }
+    });
+  });
+};
