+++ packages/commercetools/theme/helpers/filters/getFiltersForUrl.js
@@ -1,0 +1,6 @@
+export const getFiltersForUrl = (filters) => {
+  return Object.entries(filters || {}).reduce((prev, [name, filter]) => {
+    prev[name] = filter.options.filter(option => option.selected).map(option => option.value);
+    return prev;
+  }, {});
+};
