+++ packages/core/theme-module/theme/helpers/filters/applyFiltersFromUrl.js
@@ -1,0 +1,3 @@
+export const applyFiltersFromUrl = () => {
+  console.warn('Please implement this per integration');
+};
