+++ packages/core/theme-module/theme/helpers/filters/getFiltersForUrl.js
@@ -1,0 +1,4 @@
+export const getFiltersForUrl = (filters) => {
+  console.warn('Please implement this per integration');
+  return filters;
+};
