+++ packages/core/theme-module/theme/helpers/filters/index.js
@@ -1,0 +1,2 @@
+export { applyFiltersFromUrl } from './applyFiltersFromUrl';
+export { getFiltersForUrl } from './getFiltersForUrl';
